requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy<2.0  # For compatibility
plotly>=5.15.0

# Web scraping
selenium>=4.15.0
beautifulsoup4>=4.12.0

# Scheduling and automation
schedule>=1.2.0
python-dotenv>=1.0.0

# LLM integration
requests>=2.31.0
pyyaml>=6.0

# Data processing
python-dateutil>=2.8.2
pytz>=2023.3

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional for enhanced features
tqdm>=4.66.0  # Progress bars
loguru>=0.7.0  # Better logging
